Better way to handle wiping out buffers.
If the only non-floating window contains the buffer we're trying to
close, an error will occur because the floating window would be the only
one remaining. To solve that, close the floating window first, then when
the buffer is wiped out, the tab will go away, or the previous buffer
will be shown in the only tab left. It's a compromise to be sure,
because you have to reopen the BufSelect window again, but until I
figure out a better way, this will do.