Floating window (#19)
* Put the BufSelect buffer in a floating window.

This not only is a nicer look, but it also gets rid of the clunky way I
was handling/restoring the alternate buffer. I never did like that
approach.

* Better way to handle wiping out buffers.

If the only non-floating window contains the buffer we're trying to
close, an error will occur because the floating window would be the only
one remaining. To solve that, close the floating window first, then when
the buffer is wiped out, the tab will go away, or the previous buffer
will be shown in the only tab left. It's a compromise to be sure,
because you have to reopen the BufSelect window again, but until I
figure out a better way, this will do.

* Make BufSelect close itself if the user leaves the window.

* Fix: tabnew was creating extra buffers. Found a better way.

* Two spaces between columns. Make window wide enough to show CWD.

* Improved wipeout behavior, and better recursive-run detection.

BufSelect is kept open, or reopened as the case may be, when wiping out
buffers. No more need to manually open BufSelect.

The BufSelect buffer doesn't need a name. Use a script variable instead
to detect when trying to run BufSelect recursively.

* Use Unicode blocks to draw bottom border of buffer list.

* Better check for recursively running BufSelect. No extra variable.

* Update documentation in README.md.

* Combine two statements into one.

* Use the correct autocmd event for closing the BufSelect window.

WinLeave was allowing the buffer to change while leaving the window
intact. BufLeave can handle a buffer switch and close the window
instead.

* Better placement of window, centered in host window.

This helps to better identify where the selected buffer will appear. And
I did a ton of refactoring too.

* Refactor footer code using printf. Remove debugging statement.

* Move most code to an autoload module.

* Remove unnecessary variable. Just use s:Footer() instead.