Refactor the logic to close and reopen the window.
I thought I could have rewritten the code to keep the window open and
just resize or move it as needed, but this proved to be harder than
anticipated. Instead I DRYed up the logic so that the window is closed
(if it already exists) every time as part of the OpenBufSelectWindow
logic.