Fix regression error when wiping out the only window's buffer.
This happens when you have multiple tabs, and the current one has only
one window with the buffer you want to wipe out. Neovim won't let you
close that tab (which happens when wiping out its buffer) if a floating
is open. So to fix that I needed to partially reinstate the code I
refactored in an earlier commit, in the CloseBuffer function.