Preview Keys and Better Help (#22)
* Refactor the logic to close and reopen the window.

I thought I could have rewritten the code to keep the window open and
just resize or move it as needed, but this proved to be harder than
anticipated. Instead I DRYed up the logic so that the window is closed
(if it already exists) every time as part of the OpenBufSelectWindow
logic.

* Remove unnecessary function calls. Size is already set.

* Add preview mode mappings for o,s,v, and t to keep BufSelect open.

* Put the help text inside the BufSelect window.

Doing this keeps the information in front of the user when it may be
needed most. It's also a bit more polished.

* Make delete commands work silently.

* Use the literal dictionary syntax

* Make preview-mode key mappings configurable.

Instead of just prefixing the non-preview mode keys with a `g`, the
preview-mode key mapping can explicitly be designated. The `g` prefix is
still the default value.

* Change divider between keys and description in help text.

* Fix regression error when wiping out the only window's buffer.

This happens when you have multiple tabs, and the current one has only
one window with the buffer you want to wipe out. Neovim won't let you
close that tab (which happens when wiping out its buffer) if a floating
is open. So to fix that I needed to partially reinstate the code I
refactored in an earlier commit, in the CloseBuffer function.

* Fix regression error in SelectByNumber, and improve its logic.

* The colon before 'call' in the nnoremap statement was accidentally
  removed.

* Improved the buffer number search. It will first look for 'X:', and if
  not found, it will look for 'X\d*:'. This is also a visual
  improvement, much like incremental searching.

* Removed the 'normal' commands, using 'call cursor()' instead of 'gg'.

* Remove some printf function calls. Shorten CursorMoved autocmd.

* Use printf() only when field widths are needed; otherwise, just use
  string concatenation.

* Shorten the logic in the CursorMoved auto command. I just can't leave
  well enough alone.

* Reorder some functions and combine two lines into one.