Bug fix:... Solution: Change the way settings are specified. (#34)
* Bug fix:... Solution: Change the way settings are specified.

This commit fixes a major bug. If the settings (`g:BufSelectSetup`) are
reset during an edit session, the plugin would fail when trying to
access a no-longer-present setting. BufSelect would fail to create its
own buffer, replace current buffer's contents, and set it to
unmodifiable. This was very difficult to undo, and could potentially
result in lost work/data.

In addition to the bug fix, the way settings are specified has been
improved. Instead of specifying them in a global variable, an autoload
function is called with the dictionary of settings, like so:

call bufselect#settings({'mappings':{'tab':'T'}})

* Clean up the README.